+++ README.md
@@ -3,14 +3,14 @@
 ## 도메인 구현 목록
 - [ ] 주문, 결제, 종료 선택
     - [ ] 3가지 이외의 값일 경우 예외처리
-- [ ] 테이블 구현
-    - [ ] 주문 여부
-    - [ ] 주문 음식 - 수량(map)
-- [ ] 메뉴 구현
-    - [ ] 종류
-    - [ ] 번호
-    - [ ] 이름
-    - [ ] 가격
+- [x] 테이블 구현
+    - [x] 주문 여부
+    - [x] 주문 음식 - 수량(map)
+- [x] 메뉴 구현
+    - [x] 종류
+    - [x] 번호
+    - [x] 이름
+    - [x] 가격
 - [ ] 주문
     - [ ] 메뉴에 등록된 번호가 아닐 시, 예외처리
     - [ ] 해당 메뉴의 총 합계가 99개가 넘어가면 예외처리
