--- src/main/java/domain/Table.java
@@ -1,14 +1,0 @@
-package domain;
-
-public class Table {
-    private final int number;
-
-    public Table(final int number) {
-        this.number = number;
-    }
-
-    @Override
-    public String toString() {
-        return Integer.toString(number);
-    }
-}
